Disable performance.now() on mobile
The reason for this is that performance.now() uses the document's life
to measure the amount of time elapsed. On mobile it seems this is paused
when the app loses focus or the device goes to sleep. It should be
possible to toggle performance mode by detecting the `pageshow` and
`pagehide` events and calculating the time elapsed via new Date() and
adding those to an additional `pending_elapsed` counter -- at least with the
assumption that the aforementioned events correspond to the sleep/unfocus
lifecycle events.