Show elapsed/total in document.title
To make this easier, we now use `convertTime(ms)` to create the following object: {h: 0, m: 0, s: 0, ms: 0 }. This object can then be formatted further, such as the to the '{h: 00, m: 00, s: 00, ms: 000}' format with `formatTime(obj)` used for the display.

The title itself simply uses the base format and some conditionals to minimize the space taken up in the page title.